Move components to a component folder
