+++ pages/sliderBasedUi/components/SeedSelector.js
@@ -1,0 +1,25 @@
+import { Button, Flex } from '@mantine/core'
+import { useAtom } from 'jotai'
+import { seedAtom } from '../../../src/state'
+
+export const SeedSelector = () => {
+  const [seed, setSeed] = useAtom(seedAtom)
+
+  return (
+    <>
+      <Flex w={'100%'} gap={8}>
+        <Button
+          variant="outline"
+          color="gray"
+          w={'100%'}
+          onClick={() => setSeed(seed - 1)}
+        >
+          Prev Seed
+        </Button>
+        <Button w={'100%'} onClick={() => setSeed(seed + 1)}>
+          New Seed
+        </Button>
+      </Flex>
+    </>
+  )
+}
